+++ src/org/mozilla/javascript/NativeJavaMap.java
@@ -132,7 +132,7 @@ public class NativeJavaMap extends NativeJavaObject {
             List<Object> ids = new ArrayList<>(map.size());
             for (Object key : map.keySet()) {
                 if (key instanceof Integer) {
-                    ids.add((Integer) key);
+                    ids.add(key);
                 } else {
                     ids.add(ScriptRuntime.toString(key));
                 }
